add cast datetime model

+++ database/migrations/2022_05_15_031038_create_rams_table.php
@@ -16,7 +16,7 @@ class CreateRamsTable extends Migration
         Schema::create('rams', function (Blueprint $table) {
             $table->id();
             $table->string('name')->nullable(false);
-            $table->tinyInteger('capacity')->nullable(false);
+            $table->integer('capacity')->nullable(false);
             $table->tinyInteger('status')->default(1);
             $table->timestamps();
         });
